Addressed PR comments + more renaming.
