+++ lib/core/SidetreeError.ts
