+++ lib/core/versions/0.4.0/AnchorFile.ts
@@ -3,7 +3,7 @@ import Encoder from './Encoder';
 import ErrorCode from './ErrorCode';
 import Multihash from './Multihash';
 import ProtocolParameters from './ProtocolParameters';
-import SidetreeError from '../../Error';
+import SidetreeError from '../../SidetreeError';
 
 /**
  * Class containing Anchor File related operations.
