+++ lib/core/versions/0.5.0/TransactionProcessor.ts
@@ -2,14 +2,14 @@ import AnchorFileModel from './models/AnchorFileModel';
 import AnchorFile from './AnchorFile';
 import BatchFile from './BatchFile';
 import DownloadManager from '../../DownloadManager';
+import FetchResultCode from '../../../common/FetchResultCode';
 import IOperationStore from '../../interfaces/IOperationStore';
 import ITransactionProcessor from '../../interfaces/ITransactionProcessor';
 import NamedAnchoredOperationModel from '../../models/NamedAnchoredOperationModel';
 import ProtocolParameters from './ProtocolParameters';
-import SidetreeError from '../../Error';
+import SidetreeError from '../../SidetreeError';
 import timeSpan = require('time-span');
 import TransactionModel from '../../../common/models/TransactionModel';
-import { FetchResultCode } from '../../../common/FetchResultCode';
 
 /**
  * Implementation of the `ITransactionProcessor`.
