+++ lib/core/versions/latest/Did.ts
@@ -1,7 +1,7 @@
 import Encoder from './Encoder';
 import ErrorCode from './ErrorCode';
 import Multihash from './Multihash';
-import SidetreeError from '../../Error';
+import SidetreeError from '../../SidetreeError';
 
 /**
  * Class containing reusable Sidetree DID related operations.
