+++ lib/core/versions/latest/FeeManager.ts
@@ -1,6 +1,6 @@
 import ErrorCode from './ErrorCode';
 import ProtocolParameters from './ProtocolParameters';
-import SidetreeError from '../../Error';
+import SidetreeError from '../../SidetreeError';
 
 /**
  * Encapsulates the functionality to calculate and verify the blockchain transaction fees.
