+++ lib/core/versions/latest/Multihash.ts
@@ -1,7 +1,7 @@
 import * as crypto from 'crypto';
 import Encoder from './Encoder';
 import ErrorCode from './ErrorCode';
-import SidetreeError from '../../Error';
+import SidetreeError from '../../SidetreeError';
 const multihashes = require('multihashes');
 
 /**
