+++ lib/core/versions/latest/RequestHandler.ts
@@ -7,7 +7,7 @@ import Operation from './Operation';
 import OperationType from '../../enums/OperationType';
 import ProtocolParameters from './ProtocolParameters';
 import Resolver from '../../Resolver';
-import SidetreeError from '../../Error';
+import SidetreeError from '../../SidetreeError';
 import { ResponseModel, ResponseStatus } from '../../../common/Response';
 
 /**
