+++ lib/ipfs/IpfsStorage.ts
@@ -1,6 +1,6 @@
 import * as IPFS from 'ipfs';
 import FetchResult from '../common/models/FetchResult';
-import { FetchResultCode } from '../common/FetchResultCode';
+import FetchResultCode from '../common/FetchResultCode';
 
 /**
  * Class that implements the IPFS Storage functionality.
