+++ tests/core/Blockchain.spec.ts
@@ -2,7 +2,7 @@ import Blockchain from '../../lib/core/Blockchain';
 import CoreErrorCode from '../../lib/core/CoreErrorCode';
 import ReadableStream from '../../lib/common/ReadableStream';
 import ServiceVersionModel from '../../lib/common/models/ServiceVersionModel';
-import SidetreeError from '../../lib/core/Error';
+import SidetreeError from '../../lib/core/SidetreeError';
 import SharedErrorCode from '../../lib/common/SharedErrorCode';
 import TransactionModel from '../../lib/common/models/TransactionModel';
