+++ tests/core/Cas.spec.ts
@@ -1,7 +1,7 @@
 import Cas from '../../lib/core/Cas';
+import FetchResultCode from '../../lib/common/FetchResultCode';
 import ReadableStream from '../../lib/common/ReadableStream';
 import ServiceVersionModel from '../../lib/common/models/ServiceVersionModel';
-import { FetchResultCode } from '../../lib/common/FetchResultCode';
 
 describe('Cas', async () => {
   it('should return file hash of the content written.', async () => {
